Modified monitor.py

+++ cuve/monitor.py
@@ -51,9 +51,9 @@ niveau_mini = (min(MIN_values)[0])[0]
 ratio8h = (round(niveau_maxi / niveau_mini,2)-1)*100
 
 MAX_values = rrdtool.fetch(database_file, 'MAX','-s', 'end-1w', '-e', 'now')
-niveau_maxi = (max(MAX_values[2])[0])
+niveau_maxi = (float)(max(MAX_values[2])[0])
 MIN_values = rrdtool.fetch(database_file, 'MIN','-s', 'end-1w', '-e', 'now')
-niveau_mini = (min(MIN_values)[0])[0]
+niveau_mini = (float)(min(MIN_values)[0])[0]
 ratio1w = (round(niveau_maxi / niveau_mini,2)-1)*100
